+++ nvim/lua/core/init.lua
@@ -1,2 +1,3 @@
-require("core.remap")
 require("core.set")
+require("core.remap")
+require("core.commands")
